pagina basico y 404
