+++ package.json
@@ -17,6 +17,7 @@
     "react-bootstrap": "^2.6.0",
     "react-data-table-component": "^7.5.3",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.5.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
