+++ src/paginas/404.jsx
@@ -1,0 +1,11 @@
+
+const Error404 = () => {
+    return (
+      <div style={{marginLeft:"45%", marginTop:"5%"}}>
+        <h1>Error 404</h1>
+        <p>Pagina no Encontrada</p>
+      </div>
+    );
+  };
+  
+  export default Error404;
